planner: Increase 1st block delay and block count for optimization
Now that the 1st delay is avoided on sync, increase the 1st block delay
to allow for improved initial planning. Allow the entire block buffer to
be used during the initial move instead of limiting it to 3 blocks,
which often do not contain enough time for proper planning of short
segments.

+++ lib/Marlin/Marlin/src/module/planner.h
@@ -274,7 +274,7 @@ class Planner {
                             block_buffer_nonbusy,     // Index of the first non busy block
                             block_buffer_planned,     // Index of the optimally planned block
                             block_buffer_tail;        // Index of the busy block, if any
-    static uint8_t delay_before_delivering;           // This counter delays delivery of blocks when queue becomes empty to allow the opportunity of merging blocks
+    static uint32_t delay_before_delivering;          // Initial milliseconds of delay for planner optimization
 
 
     #if ENABLED(DISTINCT_E_FACTORS)
